Complete Skeleton UI for SAQ
