+++ themes/colors.ts
@@ -1,9 +1,11 @@
-export default <Record<string, string>> {
-    red: "#EF4444",
-    green: "#10B981",
-    blue: "#3B82F6",
-    yellow: "#F59E0B",
+export default <Record<string, string>>{
+  red: "#EF4444",
+  green: "#10B981",
+  blue: "#3B82F6",
+  yellow: "#F59E0B",
 
-    headerBlue: "#93C5FD",
-    headerGreen: "#A7F3D0"
-}
+  headerBlue: "#93C5FD",
+  headerGreen: "#A7F3D0",
+
+  inputBlue: "#DBEAFE",
+};
